requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==23.1.0
black==23.11.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chromedriver-autoinstaller==0.6.3
click==8.1.7
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.7
Django==3.2.19
djangorestframework==3.14.0
flake8==6.0.0
flake8-django==1.4
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
kaitaistruct==0.10
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pyasn1==0.5.1
pycodestyle==2.10.0
pycparser==2.21
pydivert==2.1.0
pyflakes==3.0.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytz==2023.3
PyVirtualDisplay==3.0
selenium==4.16.0
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
wrapt==1.16.0
wsproto==1.2.0
zstandard==0.22.0
